step 06: separate classes
